+++ source/scal.h
@@ -11,26 +11,26 @@ namespace LCPP {
 
         SCAL() {}
 
-        void operator()(size_t n, T a, T* X, size_t incx) {
+        void operator()(int n, T a, T* X, int incx) {
             apply(n, a, X, incx);
         }
 
-        void operator()(size_t n, T a, T* X) {
+        void operator()(int n, T a, T* X) {
             apply(n, a, X, 1);
         }
 
         void operator()(T a, NUMCPP::Sequence<T>& X) {
-            size_t incx = X.increment();
+            int incx = X.increment();
             apply(X.length(), a, X.begin(), incx);
         }
 
     private:
 
-        void apply(size_t n, T a, T* X, size_t incx);
+        void apply(int n, T a, T* X, int incx);
     };
 
     template <typename T>
-    void SCAL<T>::apply(size_t n, T a, T* X, size_t incx) {
+    void SCAL<T>::apply(int n, T a, T* X, int incx) {
         // Quick return if possible
         if (n == 0 || a == 1)
             return;
